- some more safety checks in the ini-limit removal code - some code cleanup and minor bug fixes based on results from static code analysis - added naemfilter for the esp list - bsa changes are now stored automatically but delayed by up to 0.5 seconds (for performance reasons) - bugfix: buffer overrun when certain functions are called with empty file names - bugfix: reroute for the ini-limit fix was placed in the data segment - bugfix: mod list got mixed up when the mod directory was changed externally - bugfix: plugins.txt reroute on skyrim didn't work on winXP - bugfix: MO could become unresponsive if the tutorial script couldn't be interpreted

+++ src/modlist.cpp
@@ -145,7 +145,7 @@ QVariant ModList::data(const QModelIndex &modelIndex, int role) const
       } else {
         if ((m_Profile->getModPriority(modIndex) == 0) && (role == Qt::DisplayRole)) {
           return tr("min");
-        } else if ((m_Profile->getModPriority(modIndex) == m_Profile->numRegularMods() - 1) &&
+        } else if ((m_Profile->getModPriority(modIndex) == static_cast<int>(m_Profile->numRegularMods()) - 1) &&
                    (role == Qt::DisplayRole)) {
           return tr("max");
         } else {
